changed the types

+++ src/main/java/kg/attractor/jobsearch/service/RespondedApplicantService.java
@@ -3,5 +3,5 @@ package kg.attractor.jobsearch.service;
 import org.apache.catalina.User;
 
 public interface RespondedApplicantService {
-    void responseVacancies(Integer vacancyId, User user);
+    void responseVacancies(Long vacancyId, User user);
 }
